changed from uint8_t to uint64_t

+++ libspaghetti/include/spaghetti/element.h
@@ -57,12 +57,12 @@ struct EventNameChanged {
 struct EventIONameChanged {
   std::string from;
   std::string to;
-  uint8_t id;
+  uint64_t id;
   bool input;
 };
 struct EventIOTypeChanged {
   bool input;
-  uint8_t id;
+  uint64_t id;
   ValueType from;
   ValueType to;
 };
@@ -105,8 +105,8 @@ class SPAGHETTI_API Element {
     Value value{};
     ValueType type{};
     size_t id{};
-    uint8_t slot{};
-    uint8_t flags{};
+    uint64_t slot{};
+    uint64_t flags{};
     std::string name{};
   };
 
@@ -151,27 +151,27 @@ class SPAGHETTI_API Element {
   IOSockets &outputs() { return m_outputs; }
   IOSockets const &outputs() const { return m_outputs; }
 
-  bool addInput(ValueType const a_type, std::string const &a_name, uint8_t const a_flags);
-  void setInputName(uint8_t const a_input, std::string const &a_name);
+  bool addInput(ValueType const a_type, std::string const &a_name, uint64_t const a_flags);
+  void setInputName(uint64_t const a_input, std::string const &a_name);
   void removeInput();
   void clearInputs();
 
-  bool addOutput(ValueType const a_type, std::string const &a_name, uint8_t const a_flags);
-  void setOutputName(uint8_t const a_output, std::string const &a_name);
+  bool addOutput(ValueType const a_type, std::string const &a_name, uint64_t const a_flags);
+  void setOutputName(uint64_t const a_output, std::string const &a_name);
   void removeOutput();
   void clearOutputs();
 
-  void setIOName(bool const a_input, uint8_t const a_id, std::string const &a_name);
-  void setIOValueType(bool const a_input, uint8_t const a_id, ValueType const a_type);
+  void setIOName(bool const a_input, uint64_t const a_id, std::string const &a_name);
+  void setIOValueType(bool const a_input, uint64_t const a_id, ValueType const a_type);
 
-  bool connect(size_t const a_sourceId, uint8_t const a_outputId, uint8_t const a_inputId);
+  bool connect(size_t const a_sourceId, uint64_t const a_outputId, uint64_t const a_inputId);
 
-  uint8_t minInputs() const { return m_minInputs; }
-  uint8_t maxInputs() const { return m_maxInputs; }
-  uint8_t defaultNewInputFlags() const { return m_defaultNewInputFlags; }
-  uint8_t minOutputs() const { return m_minOutputs; }
-  uint8_t maxOutputs() const { return m_maxOutputs; }
-  uint8_t defaultNewOutputFlags() const { return m_defaultNewOutputFlags; }
+  uint64_t minInputs() const { return m_minInputs; }
+  uint64_t maxInputs() const { return m_maxInputs; }
+  uint64_t defaultNewInputFlags() const { return m_defaultNewInputFlags; }
+  uint64_t minOutputs() const { return m_minOutputs; }
+  uint64_t maxOutputs() const { return m_maxOutputs; }
+  uint64_t defaultNewOutputFlags() const { return m_defaultNewOutputFlags; }
 
   Package *package() const { return m_package; }
 
@@ -191,13 +191,13 @@ class SPAGHETTI_API Element {
   void handleEvent(Event const &a_event);
   virtual void onEvent(Event const &a_event) { (void)a_event; }
 
-  void setMinInputs(uint8_t const a_min);
-  void setMaxInputs(uint8_t const a_max);
-  void setDefaultNewInputFlags(uint8_t const a_flags) { m_defaultNewInputFlags = a_flags; }
+  void setMinInputs(uint64_t const a_min);
+  void setMaxInputs(uint64_t const a_max);
+  void setDefaultNewInputFlags(uint64_t const a_flags) { m_defaultNewInputFlags = a_flags; }
 
-  void setMinOutputs(uint8_t const a_min);
-  void setMaxOutputs(uint8_t const a_max);
-  void setDefaultNewOutputFlags(uint8_t const a_flags) { m_defaultNewOutputFlags = a_flags; }
+  void setMinOutputs(uint64_t const a_min);
+  void setMaxOutputs(uint64_t const a_max);
+  void setDefaultNewOutputFlags(uint64_t const a_flags) { m_defaultNewOutputFlags = a_flags; }
 
  protected:
   IOSockets m_inputs{};
@@ -212,12 +212,12 @@ class SPAGHETTI_API Element {
   vec2d m_position{};
   bool m_isIconified{};
   bool m_iconifyingHidesCentralWidget{};
-  uint8_t m_minInputs{};
-  uint8_t m_maxInputs{ std::numeric_limits<uint8_t>::max() };
-  uint8_t m_minOutputs{};
-  uint8_t m_maxOutputs{ std::numeric_limits<uint8_t>::max() };
-  uint8_t m_defaultNewInputFlags{};
-  uint8_t m_defaultNewOutputFlags{};
+  uint64_t m_minInputs{};
+  uint64_t m_maxInputs{ std::numeric_limits<uint64_t>::max() };
+  uint64_t m_minOutputs{};
+  uint64_t m_maxOutputs{ std::numeric_limits<uint64_t>::max() };
+  uint64_t m_defaultNewInputFlags{};
+  uint64_t m_defaultNewOutputFlags{};
   EventCallback m_handler{};
   void *m_node{};
 };
